Moved poll to Fun class
